[FEAT] [INTERNAL] added ability to obtain connection's server version
