+++ tests/helpers/mod.ts
@@ -1,5 +1,9 @@
 import { NatsServer } from "./launcher.ts";
-import { NatsConnection } from "../../nats-base-client/types.ts";
+import {
+  compare,
+  NatsConnection,
+  parseSemVer,
+} from "../../nats-base-client/types.ts";
 import { cleanup } from "../jstest_util.ts";
 
 export { check } from "./check.ts";
@@ -13,30 +17,6 @@ export {
   assertThrowsErrorCode,
 } from "./asserts.ts";
 export { NatsServer, ServerSignals } from "./launcher.ts";
-
-export type SemVer = { major: number; minor: number; micro: number };
-export function parseSemVer(
-  s: string,
-): SemVer {
-  const m = s.match(/(\d+).(\d+).(\d+)/);
-  if (m) {
-    return {
-      major: parseInt(m[1]),
-      minor: parseInt(m[2]),
-      micro: parseInt(m[3]),
-    };
-  }
-  throw new Error(`${s} is not a semver value`);
-}
-export function compare(a: SemVer, b: SemVer): number {
-  if (a.major < b.major) return -1;
-  if (a.major > b.major) return 1;
-  if (a.minor < b.minor) return -1;
-  if (a.minor > b.minor) return 1;
-  if (a.micro < b.micro) return -1;
-  if (a.micro > b.micro) return 1;
-  return 0;
-}
 
 export function disabled(reason: string): void {
   const m = new TextEncoder().encode(red(`skipping: ${reason} `));
