+++ tests/semver_test.ts
@@ -1,4 +1,4 @@
-import { compare, parseSemVer } from "./mod.ts";
+import { compare, parseSemVer } from "../nats-base-client/types.ts";
 import { assertEquals } from "https://deno.land/std@0.95.0/testing/asserts.ts";
 
 Deno.test("semver", () => {
